refactoring : firebase
