+++ src/app/providers/store/store.tsx
@@ -2,7 +2,7 @@ import { configureStore } from "@reduxjs/toolkit";
 import { kinopoiskApi } from "../../../shared/api/kinopoiskApi/kinopoiskApi";
 import { authReducer } from "../../../shared/reducers/Authorization";
 import { favoriteReducer } from "../../../shared/reducers/Favorites";
-import { firestoreReducer } from "../../../shared/reducers/Firestor";
+import { firestoreReducer } from "../../../shared/reducers/Profile/Firestore";
 import { searchReducer } from "../../../shared/reducers/Search";
 import { historyReducer } from "../../../shared/reducers/History/";
 import { AuthLogsMiddleware } from "./middlewares/loggerMiddleware";
