+++ src/shared/reducers/Profile/Firestore/actions/FirestorActions.ts
@@ -1,6 +1,6 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import { DocumentData, doc, getDoc, setDoc } from "firebase/firestore";
-import { db } from "../../../../firebase";
+import { db } from "../../../../../firebase";
 
 interface User {
   email: string;
